add "only_last" and "only_best" to list_checkpoints.
BUG: Experiment/ExpRun currently have an issue where's it's not possible/easy in the model to say what nepochs is for a checkpoint. since all that is stored in a (metadata-only) Experiment, it contains the total nepochs for that experiment, which is often higher than that of the checkpoint.

make_denoise_anim:
* add --nlatents, --mult, --ngen
* --nlatents puts anim into a different mode:
  * instead of animating a single diffusion, it starts with (nlatents) random latents, and does a diffusion of each interpolated latent for (--steps) steps.

train_denoise:
* refactoring

loggers/tensorboard:
* add update_metadata_freq (default to 30) to limit the frequency of metadata.json write

+++ denoise/noisegen.py
@@ -110,9 +110,9 @@ class NoiseSchedule:
         steps = max(1, steps)
         steps = min(self.timesteps - 1, steps)
 
-        step_list = torch.linspace(self.timesteps - 1, 0, steps)
+        step_list = torch.linspace(self.timesteps - 1, 0, steps).int()
         for step in step_list:
-            out = self.gen_frame(net, inputs=out, timestep=int(step))
+            out = self.gen_frame(net, inputs=out, timestep=step)
 
         # for step in reversed(range(steps)):
         #     out = self.gen_frame(net, inputs=out, timestep=int(step * steps / self.timesteps))
